--- src/main/java/roomescape/controller/dto/request/ReservationRequest.java
@@ -1,13 +1,0 @@
-package roomescape.controller.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import java.time.LocalDate;
-
-public record ReservationRequest(
-        @NotBlank String name,
-        @NotNull LocalDate date,
-        @NotNull long timeId,
-        @NotNull long themeId
-) {
-}
