--- src/main/java/roomescape/controller/dto/request/ThemeCreateRequest.java
@@ -1,12 +1,0 @@
-package roomescape.controller.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-public record ThemeCreateRequest(
-        @NotBlank String name,
-        @NotNull String description,
-        @NotNull String page,
-        @NotNull String thumbnail
-) {
-}
