--- src/main/java/roomescape/controller/page/UserPageController.java
@@ -1,18 +1,0 @@
-package roomescape.controller.page;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class UserPageController {
-
-    @GetMapping("/")
-    public String getPopularPage() {
-        return "index";
-    }
-
-    @GetMapping("/reservation")
-    public String getReservationPage() {
-        return "reservation";
-    }
-}
