--- src/main/java/roomescape/model/repository/ThemeRepository.java
@@ -1,21 +1,0 @@
-package roomescape.model.repository;
-
-import roomescape.model.entity.Theme;
-
-import java.time.LocalDate;
-import java.util.List;
-
-public interface ThemeRepository {
-
-    Theme save(Theme theme);
-
-    Theme findById(long id);
-
-    List<Theme> findAll();
-
-    boolean existById(long id);
-
-    List<Theme> findPopularThemes(LocalDate startInclusive, LocalDate endInclusive, int count);
-
-    void deleteById(long id);
-}
