+++ src/main/java/roomescape/presentation/dto/request/ThemeCreateRequest.java
@@ -1,0 +1,12 @@
+package roomescape.presentation.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
+public record ThemeCreateRequest(
+        @NotBlank String name,
+        @NotNull String description,
+        @NotNull String page,
+        @NotNull String thumbnail
+) {
+}
