+++ src/main/java/roomescape/presentation/dto/response/ThemeResponse.java
@@ -1,0 +1,21 @@
+package roomescape.presentation.dto.response;
+
+import java.util.List;
+import roomescape.business.model.entity.Theme;
+
+public record ThemeResponse(
+        long id,
+        String name,
+        String description,
+        String thumbnail
+) {
+    public static ThemeResponse from(Theme theme) {
+        return new ThemeResponse(theme.getId(), theme.getName(), theme.getDescription(), theme.getThumbnail());
+    }
+
+    public static List<ThemeResponse> from(List<Theme> theme) {
+        return theme.stream()
+                .map(ThemeResponse::from)
+                .toList();
+    }
+}
