+++ src/main/java/roomescape/presentation/page/AdminPageController.java
@@ -1,0 +1,28 @@
+package roomescape.presentation.page;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AdminPageController {
+
+    @GetMapping("/admin")
+    public String getMainPage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String getReservationPage() {
+        return "admin/reservation";
+    }
+
+    @GetMapping("/admin/time")
+    public String getReservationTimePage() {
+        return "admin/time";
+    }
+
+    @GetMapping("/admin/theme")
+    public String getThemePage() {
+        return "admin/theme";
+    }
+}
