+++ src/main/java/roomescape/presentation/page/UserPageController.java
@@ -1,0 +1,18 @@
+package roomescape.presentation.page;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class UserPageController {
+
+    @GetMapping("/")
+    public String getPopularPage() {
+        return "index";
+    }
+
+    @GetMapping("/reservation")
+    public String getReservationPage() {
+        return "reservation";
+    }
+}
